# TreasureDeckApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TreasureDeck API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v2/expansions": {
            "get": {
                "tags": [
                    "expansions"
                ],
                "summary": "Listar todas las expansiones",
                "operationId": "c2dbf6b8fb4b2ed7dc15135a6748c173",
                "responses": {
                    "200": {
                        "description": "Lista de expansiones"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "expansions"
                ],
                "summary": "Crear una nueva expansión",
                "operationId": "b5d3f86a79eedbe41c96965d32e1fdd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre de expansión"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expansión creada con éxito"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/expansions/{id}": {
            "get": {
                "tags": [
                    "expansions"
                ],
                "summary": "Obtener una expansión por ID",
                "operationId": "9cd7f8843ee3b4aac7bf17374faad22b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la expansión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la expansión"
                    },
                    "404": {
                        "description": "Expansión no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "expansions"
                ],
                "summary": "Actualizar una expansión existente",
                "operationId": "fa5d61b06cf37637b68722eb2823d57f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la expansión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre de expansión"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expansión actualizada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "expansions"
                ],
                "summary": "Eliminar una expansión",
                "operationId": "b6580d9da6a66efc194b55bc04b73ed9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la expansión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expansión eliminada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/role": {
            "get": {
                "tags": [
                    "role"
                ],
                "summary": "Listar todos los role",
                "operationId": "aeb1127751e6add756471343f2bd8033",
                "responses": {
                    "200": {
                        "description": "Lista de role"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "role"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "a03fec6c5318c865f5aa5b74ae62499b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado con éxito"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/role/{id}": {
            "get": {
                "tags": [
                    "role"
                ],
                "summary": "Obtener un rol por ID",
                "operationId": "8d9d9f64a47a83cf57ad159d6fba2555",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del rol"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "role"
                ],
                "summary": "Actualizar un rol existente",
                "operationId": "257b439759845bd011a4ed03a374bdaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Editor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "role"
                ],
                "summary": "Eliminar un rol",
                "operationId": "87192da30e7248b46bd4c95cc5dbc82e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Rol eliminado con éxito"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Autenticación usando Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}